Updated the duration field to float

+++ Video/Interfaces/IMedia.cs
@@ -8,7 +8,7 @@ namespace Gorilla.Storage.Video.Interfaces
         string Id { get; set; }
         string Name { get; set; }
         string Description { get; set; }
-        int Duration { get; set; }
+        float Duration { get; set; }
         UploadStatus Status { get; set; }
         float Progress { get; set; }
         string Thumbnail { get; set; }
